feat: :sparkles: Make project work with GCP Cloud Functions

+++ src/services/discord.service.ts
@@ -111,6 +111,9 @@ export default class DiscordService {
   }
 
   disconnect(): void {
+    if (!this.client) {
+      return;
+    }
     this.client.destroy();
   }
 }
